Port filter tests to test_dynamic_api (AB#24838)
These too need to use a schema in the new setup.
